+++ src/inference/metropolis_hastings.cpp
@@ -3,7 +3,7 @@
 #include "inference/metropolis_sampler.hpp"
 #include "multivariate_normal.hpp"
 
-MetropolisHastings::MetropolisHastings(std::function<double(Eigen::VectorXd)>& posterior, double scale, int nsims,
+MetropolisHastings::MetropolisHastings(std::function<double(Eigen::VectorXd)>& posterior, double scale, size_t nsims,
                                        const Eigen::VectorXd& initials,
                                        const std::optional<Eigen::MatrixXd>& cov_matrix, int thinning,
                                        bool warm_up_period, // TODO: TSM model_object = nullptr,
